pyproject.toml
[project]
name = "dist-sys"
version = "0.1.0"
description = "Parallel Conway's Game of Life using MPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mpi4py>=3.1.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "scipy>=1.11.0",
]
